% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.marketplaceentitlementservice_package.R
\docType{package}
\name{paws.marketplaceentitlementservice-package}
\alias{paws.marketplaceentitlementservice}
\alias{paws.marketplaceentitlementservice-package}
\title{paws.marketplaceentitlementservice: AWS Marketplace Entitlement Service}
\description{
This reference provides descriptions of the AWS
Marketplace Entitlement Service API.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
